test(storage): fold factory-chain coverage into the existing capability suite
Valid: the new file restated concerns `managed-read-gate-decorator-coverage-v1`
already owns. Its per-decorator assertions duplicated "resolves through every
storage decorator" and "every first-party storage wrapper is traversable by
REGISTRATION alone", and its negative control duplicated "returns null for a
store with no managed backend anywhere".

Exactly one property in it was new: that the composition `createTripleStore`
actually builds resolves end to end. That test moves into the existing suite and
the file is deleted, along with its workflow entry.

Worth recording what this leaves. This PR opened claiming mechanical enforcement
for future decorators and ends as one added test asserting one property. Each
round removed something that was not what it claimed: a heuristic selector
reported as enforcement, a tripwire that asserted a constant against itself, a
hand-rolled traversal inside a suite about traversal, and now a duplicate model
of coverage that already existed. What survives is the part that was actually
missing.

Still discriminating: inserting a fourth non-participating decorator into
`createTripleStore` reddens the merged test; removing it goes green. Storage
lane 30 files / 525 tests.
